requirements.txt
django==2.0.9
numpy==1.15.1
pandas==0.23.4
psycopg2-binary==2.7.7
django-crispy-forms==1.7.2